+++ src/views/scroll-editor/artefact-toolbox.vue
@@ -791,17 +791,11 @@ export default class ArtefactToolbox extends Vue {
 </script>
 
 <style lang="scss" scoped>
-.delete {
-    color: red;
-}
 .btn-xs {
     padding: 0.1rem 0.15rem;
     font-size: 0.75rem;
     line-height: 1;
     border-radius: 0.2rem;
-}
-.btn-menu {
-    background: #e3e7ea;
 }
 .card-body-cancel {
     padding: 0rem !important;
